ts_db_apis/requirements.txt

fastapi
uvicorn
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic
aiofiles
requests
fastapi[all]
twilio
Pillow
pytz
argon2-cffi
email-validator
jose